backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
python-dotenv
psycopg2-binary
requests
django-async-redis
fastapi
uvicorn[standard]
httpx
pytest
pytest-asyncio
apscheduler